Incident timeline chart, a real anomaly detector, and one writer for day/night
The fourth chart needed incidents and the table was empty. Rather than generate
plausible ones, two honest sources: a residual anomaly detector, and the 18,578
crash records already in the warehouse.

The detector scores each link-bucket against that link's own median for its
weekday-hour, using MAD as the scale. Both statistics are robust on purpose — a
mean and a standard deviation are pulled up by the incidents you are looking
for, so a naive z-score detector finds fewer the more there are. Two tests must
pass, statistically unusual and materially worse, because a near-constant link
has a near-zero MAD that turns a two-point wobble into a huge z-score. Over the
seeded day it finds three low-severity anomalies, which is reported as three:
the seed is a smooth profile and nothing put anomalies into it. A satisfying
number here would have meant a broken detector.

The chart itself is five years of crashes by hour, banded by injury outcome,
with the congestion curve over it. They peak at the same hour, 18:00. That is
the whole road-safety argument in one panel.

Light mode was rendering a night city inside a white interface because two
components wrote data-scene: the header toggle, and the 3D view's own useState
whose effect re-asserted night every render. The console shells then passed
scene="night" as a literal. lib/theme.ts is now the only writer and the store is
the document itself, so there is no cached copy to drift. The daylight scene
already existed in full — it had simply never been told the time of day.

Resting zoom moves from fit*0.028 to fit*0.12. The close range was tuned for a
one-off "show me the vehicles" and had quietly become the view the dashboard
opens to: two cars and no city. A close-up is somewhere you zoom to.

Also: `AS index` read back as `r.index` shadows tuple.index. It resolved to the
column at runtime, which is why it survived, but it reads as the method and
mypy types it as one. Renamed, and mypy is clean across the API and adapters
for the first time.

+++ packages/adapters/src/pravaah/adapters/profiles.py
@@ -198,7 +198,7 @@ def speed_kmh(index: float, free_flow_kmh: float) -> float:
     the peak minute, which is what that figure is often misread as.
     """
     ratio = 1.0 - _SPEED_DECAY_COEFFICIENT * (max(0.0, index) / 100.0) ** _SPEED_DECAY_EXPONENT
-    return round(free_flow_kmh * max(_MIN_SPEED_RATIO, ratio), 1)
+    return float(round(free_flow_kmh * max(_MIN_SPEED_RATIO, ratio), 1))
 
 
 def rush_hour_mean_speed(free_flow_kmh: float = CALIBRATION_FREE_FLOW_KMH) -> float:
